items: move storage to TRX
This moves item storage to TRX, along with some functions that directly
manipulate the array such as killing and adding active items.
