+++ src/libtrx/include/libtrx/game/items/const.h
@@ -1,3 +1,8 @@
 #pragma once
 
 #define NO_ITEM (-1)
+#if TR_VERSION == 1
+    #define MAX_ITEMS 10240
+#else
+    #define MAX_ITEMS 256
+#endif
