+++ src/tr1/game/items.h
@@ -4,22 +4,12 @@
 
 #include <stdint.h>
 
-extern ITEM *g_Items;
-
-void Item_InitialiseArray(int32_t num_items);
 void Item_Control(void);
-void Item_Kill(int16_t item_num);
 void Item_Initialise(int16_t item_num);
-void Item_RemoveActive(int16_t item_num);
-void Item_RemoveDrawn(int16_t item_num);
-void Item_AddActive(int16_t item_num);
-void Item_NewRoom(int16_t item_num, int16_t room_num);
 void Item_UpdateRoom(ITEM *item, int32_t height);
 int16_t Item_GetHeight(ITEM *item);
 int16_t Item_GetWaterHeight(ITEM *item);
 int16_t Item_Spawn(const ITEM *item, GAME_OBJECT_ID obj_id);
-int32_t Item_GlobalReplace(
-    GAME_OBJECT_ID src_obj_id, GAME_OBJECT_ID dst_obj_id);
 
 bool Item_IsNearItem(const ITEM *item, const XYZ_32 *pos, int32_t distance);
 bool Item_Test3DRange(int32_t x, int32_t y, int32_t z, int32_t range);
