+++ src/tr1/global/const.h
@@ -11,7 +11,6 @@
 #define MAX_FRAMES 10
 #define MAX_CD_TRACKS 64
 #define MAX_FLIP_MAPS 10
-#define MAX_ITEMS 10240
 #define MAX_SECRETS 16
 #define LARA_MAX_HITPOINTS 1000
 #define LARA_MAX_AIR 1800
