+++ src/tr2/game/items.h
@@ -2,16 +2,8 @@
 
 #include "global/types.h"
 
-void Item_InitialiseArray(int32_t num_items);
 void Item_Control(void);
-void Item_Kill(int16_t item_num);
 void Item_Initialise(int16_t item_num);
-void Item_RemoveActive(int16_t item_num);
-void Item_RemoveDrawn(int16_t item_num);
-void Item_AddActive(int16_t item_num);
-void Item_NewRoom(int16_t item_num, int16_t room_num);
-int32_t Item_GlobalReplace(
-    GAME_OBJECT_ID src_obj_id, GAME_OBJECT_ID dst_obj_id);
 void Item_ClearKilled(void);
 void Item_ShiftCol(ITEM *item, COLL_INFO *coll);
 void Item_UpdateRoom(ITEM *item, int32_t height);
